nsqd: remove leftover go1.0 compatible files
