+++ nsqd/buffer_pool.go
@@ -1,5 +1,3 @@
-// +build go1.3
-
 package nsqd
 
 import (
