--- nsqd/buffer_pool_go10.go
@@ -1,13 +1,0 @@
-// +build !go1.3
-
-package nsqd
-
-import (
-	"bytes"
-)
-
-func bufferPoolGet() *bytes.Buffer {
-	return &bytes.Buffer{}
-}
-
-func bufferPoolPut(b *bytes.Buffer) {}
