--- nsqd/client_tls_go10.go
@@ -1,7 +1,0 @@
-// +build !go1.3
-
-package nsqd
-
-func (p *prettyConnectionState) GetVersion() string {
-	return "TLS1.0"
-}
