+++ nsqd/client_v2.go
@@ -267,54 +267,51 @@ type prettyConnectionState struct {
 	tls.ConnectionState
 }
 
-// taken from http://golang.org/src/pkg/crypto/tls/cipher_suites.go
-// to be compatible with older versions
-const (
-	local_TLS_RSA_WITH_RC4_128_SHA                uint16 = 0x0005
-	local_TLS_RSA_WITH_3DES_EDE_CBC_SHA           uint16 = 0x000a
-	local_TLS_RSA_WITH_AES_128_CBC_SHA            uint16 = 0x002f
-	local_TLS_RSA_WITH_AES_256_CBC_SHA            uint16 = 0x0035
-	local_TLS_ECDHE_ECDSA_WITH_RC4_128_SHA        uint16 = 0xc007
-	local_TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA    uint16 = 0xc009
-	local_TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA    uint16 = 0xc00a
-	local_TLS_ECDHE_RSA_WITH_RC4_128_SHA          uint16 = 0xc011
-	local_TLS_ECDHE_RSA_WITH_3DES_EDE_CBC_SHA     uint16 = 0xc012
-	local_TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA      uint16 = 0xc013
-	local_TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA      uint16 = 0xc014
-	local_TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256   uint16 = 0xc02f
-	local_TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256 uint16 = 0xc02b
-)
-
 func (p *prettyConnectionState) GetCipherSuite() string {
 	switch p.CipherSuite {
-	case local_TLS_RSA_WITH_RC4_128_SHA:
+	case tls.TLS_RSA_WITH_RC4_128_SHA:
 		return "TLS_RSA_WITH_RC4_128_SHA"
-	case local_TLS_RSA_WITH_3DES_EDE_CBC_SHA:
+	case tls.TLS_RSA_WITH_3DES_EDE_CBC_SHA:
 		return "TLS_RSA_WITH_3DES_EDE_CBC_SHA"
-	case local_TLS_RSA_WITH_AES_128_CBC_SHA:
+	case tls.TLS_RSA_WITH_AES_128_CBC_SHA:
 		return "TLS_RSA_WITH_AES_128_CBC_SHA"
-	case local_TLS_RSA_WITH_AES_256_CBC_SHA:
+	case tls.TLS_RSA_WITH_AES_256_CBC_SHA:
 		return "TLS_RSA_WITH_AES_256_CBC_SHA"
-	case local_TLS_ECDHE_ECDSA_WITH_RC4_128_SHA:
+	case tls.TLS_ECDHE_ECDSA_WITH_RC4_128_SHA:
 		return "TLS_ECDHE_ECDSA_WITH_RC4_128_SHA"
-	case local_TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA:
+	case tls.TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA:
 		return "TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA"
-	case local_TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA:
+	case tls.TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA:
 		return "TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA"
-	case local_TLS_ECDHE_RSA_WITH_RC4_128_SHA:
+	case tls.TLS_ECDHE_RSA_WITH_RC4_128_SHA:
 		return "TLS_ECDHE_RSA_WITH_RC4_128_SHA"
-	case local_TLS_ECDHE_RSA_WITH_3DES_EDE_CBC_SHA:
+	case tls.TLS_ECDHE_RSA_WITH_3DES_EDE_CBC_SHA:
 		return "TLS_ECDHE_RSA_WITH_3DES_EDE_CBC_SHA"
-	case local_TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA:
+	case tls.TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA:
 		return "TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA"
-	case local_TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA:
+	case tls.TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA:
 		return "TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA"
-	case local_TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256:
+	case tls.TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256:
 		return "TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256"
-	case local_TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256:
+	case tls.TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256:
 		return "TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256"
 	}
-	return fmt.Sprintf("unkown %d", p.CipherSuite)
+	return fmt.Sprintf("Unknown %d", p.CipherSuite)
+}
+
+func (p *prettyConnectionState) GetVersion() string {
+	switch p.Version {
+	case tls.VersionSSL30:
+		return "SSL30"
+	case tls.VersionTLS10:
+		return "TLS1.0"
+	case tls.VersionTLS11:
+		return "TLS1.1"
+	case tls.VersionTLS12:
+		return "TLS1.2"
+	default:
+		return fmt.Sprintf("Unknown %d", p.Version)
+	}
 }
 
 func (c *clientV2) IsReadyForMessages() bool {
